Refactor session replay service and update imports
- Replaced instances of `SessionReplayService` with `SessionReplayQueryService` and `SessionReplayIngestService` in relevant API files to improve clarity and functionality.
- Removed the obsolete `trackingUtils.ts` file and the `sessionReplayService.ts` file to streamline the codebase.

Files modified:
- `server/src/api/sessionReplay/getSessionReplayEvents.ts`
- `server/src/api/sessionReplay/getSessionReplays.ts`
- `server/src/api/sessionReplay/markSessionComplete.ts`
- `server/src/api/sessionReplay/recordSessionReplay.ts`
- `server/src/lib/trackingUtils.ts`
- `server/src/services/sessionReplayService.ts`
