+++ server/src/services/sessionReplayIngestService.ts
@@ -1,0 +1,200 @@
+import { clickhouse } from "../db/clickhouse/clickhouse.js";
+import { RecordSessionReplayRequest } from "../types/sessionReplay.js";
+import { processResults } from "../api/analytics/utils.js";
+import { parseTrackingData } from "./trackingUtils.js";
+
+export interface RequestMetadata {
+  userAgent: string;
+  ipAddress: string;
+  origin: string;
+  referrer: string;
+}
+
+/**
+ * Service responsible for ingesting session replay data
+ * Handles recording events and updating metadata
+ */
+export class SessionReplayIngestService {
+  async recordEvents(
+    siteId: number,
+    request: RecordSessionReplayRequest,
+    requestMeta?: RequestMetadata
+  ): Promise<void> {
+    const { sessionId, userId, events, metadata } = request;
+
+    // Prepare events for batch insert
+    const eventsToInsert = events.map((event, index) => ({
+      site_id: siteId,
+      session_id: sessionId,
+      user_id: userId,
+      timestamp: Math.floor(event.timestamp / 1000), // Convert to Unix timestamp (seconds)
+      event_type: event.type,
+      event_data: JSON.stringify(event.data),
+      sequence_number: index,
+      event_size_bytes: JSON.stringify(event.data).length,
+      viewport_width: metadata?.viewportWidth || null,
+      viewport_height: metadata?.viewportHeight || null,
+      is_complete: 0,
+    }));
+
+    // Batch insert events
+    if (eventsToInsert.length > 0) {
+      await clickhouse.insert({
+        table: "session_replay_events",
+        values: eventsToInsert,
+        format: "JSONEachRow",
+      });
+    }
+
+    // Update or insert metadata
+    if (metadata) {
+      await this.updateSessionMetadata(
+        siteId,
+        sessionId,
+        userId,
+        metadata,
+        requestMeta
+      );
+    }
+  }
+
+  private async updateSessionMetadata(
+    siteId: number,
+    sessionId: string,
+    userId: string,
+    metadata: any,
+    requestMeta?: RequestMetadata
+  ): Promise<void> {
+    // Get existing session info from events table
+    const sessionInfo = await clickhouse.query({
+      query: `
+        SELECT 
+          MIN(timestamp) as start_time,
+          MAX(timestamp) as end_time,
+          COUNT() as event_count,
+          SUM(event_size_bytes) as compressed_size_bytes,
+          MAX(viewport_width) as screen_width,
+          MAX(viewport_height) as screen_height
+        FROM session_replay_events
+        WHERE site_id = {siteId:UInt16} 
+          AND session_id = {sessionId:String}
+      `,
+      query_params: { siteId, sessionId },
+      format: "JSONEachRow",
+    });
+
+    type SessionInfoResult = {
+      start_time: string;
+      end_time: string | null;
+      event_count: number;
+      compressed_size_bytes: number;
+      screen_width: number | null;
+      screen_height: number | null;
+    };
+
+    const sessionResults = await processResults<SessionInfoResult>(sessionInfo);
+
+    if (!sessionResults || sessionResults.length === 0) return;
+
+    const sessionReplayData = sessionResults[0];
+
+    // Parse tracking data from request metadata
+    let trackingData: any = {};
+    if (requestMeta?.userAgent) {
+      try {
+        // Extract hostname from the page URL
+        const urlObj = new URL(metadata.pageUrl);
+        const hostname = urlObj.hostname;
+
+        trackingData = await parseTrackingData(
+          requestMeta.userAgent,
+          requestMeta.ipAddress,
+          requestMeta.referrer || "",
+          urlObj.search || "", // querystring from URL
+          hostname,
+          metadata.language || "", // language from client
+          sessionReplayData.screen_width || metadata.viewportWidth || 0,
+          sessionReplayData.screen_height || metadata.viewportHeight || 0
+        );
+      } catch (error) {
+        console.error("Error parsing tracking data for session replay:", error);
+      }
+    }
+
+    // Calculate duration
+    const startTime = new Date(sessionReplayData.start_time);
+    const endTime = sessionReplayData.end_time
+      ? new Date(sessionReplayData.end_time)
+      : null;
+    const durationMs = endTime ? endTime.getTime() - startTime.getTime() : null;
+
+    // Insert or update metadata
+    await clickhouse.insert({
+      table: "session_replay_metadata",
+      values: [
+        {
+          site_id: siteId,
+          session_id: sessionId,
+          user_id: userId,
+          start_time: Math.floor(startTime.getTime() / 1000), // Convert to Unix timestamp
+          end_time: endTime ? Math.floor(endTime.getTime() / 1000) : null, // Convert to Unix timestamp
+          duration_ms: durationMs,
+          event_count: sessionReplayData.event_count || 0,
+          compressed_size_bytes: sessionReplayData.compressed_size_bytes || 0,
+          page_url: metadata.pageUrl || "",
+          country: trackingData.country || "",
+          region: trackingData.region || "",
+          city: trackingData.city || "",
+          lat: trackingData.lat || 0,
+          lon: trackingData.lon || 0,
+          browser: trackingData.browser || "",
+          browser_version: trackingData.browserVersion || "",
+          operating_system: trackingData.operatingSystem || "",
+          operating_system_version: trackingData.operatingSystemVersion || "",
+          language: trackingData.language || "",
+          screen_width:
+            sessionReplayData.screen_width || metadata?.viewportWidth || 0,
+          screen_height:
+            sessionReplayData.screen_height || metadata?.viewportHeight || 0,
+          device_type: trackingData.deviceType || "",
+          channel: trackingData.channel || "",
+          hostname: trackingData.hostname || "",
+          referrer: trackingData.referrer || "",
+          has_replay_data: 1,
+          recording_status: "recording",
+        },
+      ],
+      format: "JSONEachRow",
+    });
+  }
+
+  async markSessionComplete(siteId: number, sessionId: string): Promise<void> {
+    // Update recording status
+    await clickhouse.query({
+      query: `
+        ALTER TABLE session_replay_metadata 
+        UPDATE recording_status = 'completed' 
+        WHERE site_id = {siteId:UInt16} 
+          AND session_id = {sessionId:String}
+      `,
+      query_params: { siteId, sessionId },
+    });
+
+    // Mark last event as complete
+    await clickhouse.query({
+      query: `
+        ALTER TABLE session_replay_events 
+        UPDATE is_complete = 1 
+        WHERE site_id = {siteId:UInt16} 
+          AND session_id = {sessionId:String}
+          AND sequence_number = (
+            SELECT MAX(sequence_number) 
+            FROM session_replay_events 
+            WHERE site_id = {siteId:UInt16} 
+              AND session_id = {sessionId:String}
+          )
+      `,
+      query_params: { siteId, sessionId },
+    });
+  }
+}
